fix debug crashes, make debug files pep8 compliant
