+++ scripts/debug_commands/cat.py
@@ -1,8 +1,8 @@
 from typing import List
 
-from scripts.debugCommands.command import Command
+from scripts.debug_commands.command import Command
 
-from scripts.debugCommands.utils import add_output_line_to_log
+from scripts.debug_commands.utils import add_output_line_to_log
 
 from scripts.game_structure.game_essentials import game
 from scripts.cat.cats import Cat
