+++ scripts/debug_commands/command.py
