+++ scripts/debug_commands/eval.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from scripts.debugCommands.command import Command
-from scripts.debugCommands.utils import add_output_line_to_log, add_multiple_lines_to_log
+from scripts.debug_commands.command import Command
+from scripts.debug_commands.utils import add_output_line_to_log, add_multiple_lines_to_log
 
 import builtins
 from typing import List
@@ -20,7 +20,7 @@ class EvalCommand(Command):
         if len(args) == 0:
             add_output_line_to_log(f"Usage: {self.name} {self.usage}")
             return
-        global warningAccepted  # pylint: disable=global-statement
+        global warningAccepted  # pylint: disable=global-statement,global-variable-not-assigned
         if not warningAccepted:
             add_multiple_lines_to_log("""WARNING: This command can be used to run code in your game. Only use this if you know what you're doing.
                                          If you have been told to use this by anyone other than the official clangen discord contributors, BLOCK THEM IMMEDIATELY.
