+++ scripts/debug_commands/fps.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from scripts.debugCommands.command import Command
-from scripts.debugCommands.utils import add_output_line_to_log
+from scripts.debug_commands.command import Command
+from scripts.debug_commands.utils import add_output_line_to_log
 from typing import List
 
 from scripts.game_structure.game_essentials import game
@@ -14,10 +14,11 @@ class FpsCommand(Command):
 
     def callback(self, args: List[str]):
         if len(args) == 1:
-            if args[0].lower() in ['none', 'null', '0'] or args[0].is_integer() and int(args[0]) <= 0:
+            if args[0].lower() in ['none', 'null', 'false'] or (args[0].lstrip('-').isnumeric() and int(args[0]) <= 0):
+                # weird workaround to allow 0 and negative numbers to remove the FPS cap
                 game.switches['fps'] = 0
                 add_output_line_to_log("FPS cap removed")
-            elif args[0].is_integer():
+            elif args[0].isnumeric():
                 game.switches['fps'] = int(args[0])
                 add_output_line_to_log(f"FPS cap set to {args[0]}")
             else:
