+++ scripts/debug_commands/utils.py
